Move files back to original format
