+++ stag_lib/graph.h
