+++ stag_lib/graphio.cpp
