+++ stag_lib/kde.cpp
@@ -5,7 +5,7 @@
 
 #include "kde.h"
 
-#include "Graph/random.h"
+#include "random.h"
 #include "multithreading/ctpl_stl.h"
 #include "data.h"
 #include <iostream>
