+++ stag_lib/kde.h
