+++ stag_lib/lsh.h
