+++ stag_lib/spectrum.cpp
